Revert recent "git merge <msg> HEAD <commit>..." deprecation
This reverts commit d2a8f0bfea6158ab347a910e4fba0f9159c7e17a "git-pull.sh:
Fix call to git-merge for new command format" and

commit 4703d18dafe57e15848e298ce7d0fb2dad316fb5 "git-merge: a deprecation
notice of the ancient command line syntax".

They caused a "git pull" (without any arguments, and without any local
commits---only to update to the other side) to warn that commit log
message is ignored because the merge resulted in a fast-forward.

Another possible solution is to add an extra option to "git merge" so that
"git pull" can tell it that the message given is not coming from the end
user (the canned message is passed just in case the merge resulted in a
non-ff and caused commit), but I think it is easier _not_ to deprecate the
old syntax.

Signed-off-by: Junio C Hamano <gitster@pobox.com>